+++ src/dot/error.rs
@@ -18,10 +18,6 @@ pub enum Error {
   #[diagnostic(code(dotfiles::walk))]
   WalkingDotfiles(#[source] walkdir::Error),
 
-  #[error("The \"dots.{0}\" file is deprecated and support will be removed in a future version.")]
-  #[diagnostic(code(dot::defaults::deprecated), severity(warning), help("Please rename this file to \"defaults.{0}\"."))]
-  DotsDeprecated(String, #[label] SourceSpan, #[source_code] String),
-
   #[cfg(feature = "yaml")]
   #[error("Could not parse dot")]
   #[diagnostic(code(dot::parse))]